+++ Magna/src/main/java/de/morigm/magna/edits/RegisterAutoEdits.java
@@ -8,6 +8,8 @@ import de.morigm.magna.api.autoedit.PlayerAutoEditStruct;
 import de.morigm.magna.edits.player.AutoEdit_PlayerName;
 import de.morigm.magna.edits.player.AutoEdit_PlayerUUID;
 import de.morigm.magna.edits.player.AutoEdit_WorldName;
+import de.morigm.magna.edits.player.AutoEdit_WorldPlayerCount;
+import de.morigm.magna.edits.server.AutoEdit_Motd;
 import de.morigm.magna.edits.server.AutoEdit_PlayerCount;
 import lombok.Getter;
 
@@ -20,6 +22,7 @@ public class RegisterAutoEdits
 	public void registerServerStruct()
 	{
 		new AutoEdit_PlayerCount().register();
+		new AutoEdit_Motd().register();
 	}
 
 	public void registerPlayerStruct() 
@@ -27,6 +30,7 @@ public class RegisterAutoEdits
 		new AutoEdit_WorldName().register();
 		new AutoEdit_PlayerName().register();
 		new AutoEdit_PlayerUUID().register();
+		new AutoEdit_WorldPlayerCount().register();
 	}
 
 }